extensible refactor
