--- index.html
@@ -1,46 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Sitemap Explorer - Input</title>
-    <link rel="stylesheet" href="styles.css">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0-beta3/css/all.min.css">
-</head>
-<body>
-    <header>
-        <h1>Sitemap Explorer <span class="version">v1.0</span></h1>
-        <div class="header-controls">
-            <button id="themeToggleBtn" aria-label="Toggle dark mode"><i class="fas fa-moon"></i></button>
-        </div>
-    </header>
-    <main>
-        <section id="controls">
-            <h2>Load Sitemap</h2>
-            <form id="urlForm" method="POST" action="/sitemap-url">
-                <div class="input-group">
-                    <input type="text" id="sitemapUrl" name="url" placeholder="e.g., https://example.com/sitemap.xml" aria-label="Sitemap URL" aria-describedby="urlTip">
-                    <span id="urlTip" class="tooltip">Enter a valid sitemap URL</span>
-                    <button type="submit" id="loadUrlBtn"><i class="fas fa-globe"></i> Load URL</button>
-                </div>
-            </form>
-            <form id="uploadForm" method="POST" action="/sitemap-upload" enctype="multipart/form-data">
-                <div class="input-group">
-                    <input type="file" id="sitemapFile" name="sitemapFile" accept=".xml" aria-label="Upload sitemap file">
-                    <button type="submit" id="uploadFileBtn"><i class="fas fa-upload"></i> Upload File</button>
-                </div>
-            </form>
-            <div class="button-group">
-                <button id="clearCacheBtn"><i class="fas fa-trash"></i> Clear Cache</button>
-            </div>
-            <div id="status">
-                <div id="loading" class="hidden">Loading... <progress id="loadProgress" value="0" max="100"></progress></div>
-                <div id="feedback" class="hidden"></div>
-                <div id="error" class="hidden"></div>
-            </div>
-        </section>
-    </main>
-    <footer></footer>
-    <script src="input.js"></script>
-</body>
-</html>
