+++ package.json
@@ -1,8 +1,10 @@
 {
   "name": "sitemap-explorer",
   "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
+  "type": "module",
+  "scripts": {
+    "start": "node src/server/index.js"
+  },
   "dependencies": {
     "axios": "^1.8.4",
     "axios-retry": "^4.5.0",
@@ -11,12 +13,5 @@
     "multer": "^1.4.5-lts.2",
     "uuid": "^10.0.0",
     "xmldom": "^0.6.0"
-  },
-  "devDependencies": {},
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC"
+  }
 }
