+++ public/index.html
@@ -1,0 +1,32 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Sitemap Explorer</title>
+    <link rel="stylesheet" href="/client/styles/styles.css">
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0-beta3/css/all.min.css">
+</head>
+<body>
+    <header>
+        <h1>Sitemap Explorer <span class="version">v1.0</span></h1>
+    </header>
+    <main>
+        <section>
+            <form id="sitemapUrlForm" method="POST">
+                <label for="sitemapUrlInput">Enter Sitemap URL:</label>
+                <input type="text" id="sitemapUrlInput" name="url" placeholder="https://example.com/sitemap.xml">
+                <button type="submit">Load Sitemap</button>
+            </form>
+            <form id="sitemapUploadForm" method="POST" enctype="multipart/form-data">
+                <label for="sitemapFile">Upload Sitemap File:</label>
+                <input type="file" id="sitemapFile" name="sitemapFile">
+                <button type="submit">Upload</button>
+            </form>
+            <button id="clearCacheBtn">Clear Cache</button>
+        </section>
+    </main>
+    <footer></footer>
+    <script type="module" src="/client/index.js"></script>
+</body>
+</html>
