+++ public/results.html
@@ -1,0 +1,56 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Sitemap Explorer - Results</title>
+    <link rel="stylesheet" href="/client/styles/styles.css">
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0-beta3/css/all.min.css">
+</head>
+<body>
+    <header>
+        <h1>Sitemap Explorer <span class="version">v1.0</span></h1>
+        <div class="header-controls">
+            <button id="themeToggleBtn" aria-label="Toggle dark mode"><i class="fas fa-moon"></i></button>
+        </div>
+    </header>
+    <main>
+        <section id="results">
+            <button id="backBtn"><i class="fas fa-arrow-left"></i> Back</button>
+            <div class="export-buttons">
+                <button id="exportHtmlBtn" disabled><i class="fas fa-file-code"></i> Export HTML</button>
+                <button id="exportCsvBtn" disabled><i class="fas fa-file-csv"></i> Export CSV</button>
+                <button id="exportJsonBtn" disabled><i class="fas fa-file-export"></i> Export JSON</button>
+            </div>
+            <button id="collapseAllBtn"><i class="fas fa-compress-alt"></i> Collapse All</button>
+            <label for="urlSearch">Search URLs:</label>
+            <input type="text" id="urlSearch" placeholder="Search URLs...">
+            <button id="toggleFiltersBtn" aria-expanded="false">Filter and Sort</button>
+            <div id="filters" class="hidden">
+                <label for="priorityFilter">Min Priority:</label>
+                <input type="number" id="priorityFilter" min="0" max="1" step="0.1" value="0">
+                <label for="lastmodFilter">Last Modified After:</label>
+                <input type="date" id="lastmodFilter">
+                <label for="sortBy">Sort By:</label>
+                <select id="sortBy">
+                    <option value="url">URL</option>
+                    <option value="lastmod">Last Modified</option>
+                    <option value="priority">Priority</option>
+                </select>
+                <button id="applyFiltersBtn">Apply Filters</button>
+                <button id="resetFiltersBtn">Reset Filters</button>
+            </div>
+            <div id="status">
+                <div id="loading" class="hidden">Loading... <progress id="loadProgress" value="0" max="100"></progress></div>
+                <div id="feedback" class="hidden"></div>
+                <div id="error" class="hidden"></div>
+            </div>
+            <p>Total URLs: <span id="totalUrls">0</span> <span id="matchCount"></span></p>
+            <p>Unique Domains: <span id="uniqueDomains">0</span></p>
+            <div id="tree" class="tree"></div>
+        </section>
+    </main>
+    <footer></footer>
+    <script type="module" src="/client/index.js"></script>
+</body>
+</html>
