--- results.html
@@ -1,67 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Sitemap Explorer - Results</title>
-    <link rel="stylesheet" href="styles.css">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0-beta3/css/all.min.css">
-</head>
-<body>
-    <header>
-        <h1>Sitemap Explorer <span class="version">v1.0</span></h1>
-        <div class="header-controls">
-            <button id="themeToggleBtn" aria-label="Toggle dark mode"><i class="fas fa-moon"></i></button>
-        </div>
-    </header>
-    <main>
-        <section id="controls">
-            <div class="button-group">
-                <button id="backBtn"><i class="fas fa-arrow-left"></i> Back to Input</button>
-                <button id="exportHtmlBtn" disabled><i class="fas fa-file-code"></i> Export HTML</button>
-                <button id="exportCsvBtn" disabled><i class="fas fa-file-csv"></i> Export CSV</button>
-                <button id="exportJsonBtn" disabled><i class="fas fa-file-code"></i> Export JSON</button>
-                <button id="collapseAllBtn"><i class="fas fa-compress-alt"></i> Collapse All</button>
-            </div>
-        </section>
-        <section id="tree-view">
-            <div id="search-and-filters">
-                <div id="search">
-                    <label for="urlSearch">Search URLs:</label>
-                    <input type="text" id="urlSearch" placeholder="e.g., /news/" aria-describedby="searchTip">
-                    <span id="searchTip" class="tooltip">Search within loaded sitemap</span>
-                </div>
-                <button id="toggleFiltersBtn" aria-expanded="false">Advanced Filters <i class="fas fa-caret-down"></i></button>
-                <div id="filters" class="hidden">
-                    <div id="stats">
-                        <p>Total URLs: <span id="totalUrls">0</span></p>
-                        <p>Unique Domains: <span id="uniqueDomains">0</span></p>
-                    </div>
-                    <label for="priorityFilter">Min Priority:</label>
-                    <input type="number" id="priorityFilter" min="0" max="1" step="0.1" value="0" aria-describedby="priorityTip">
-                    <span id="priorityTip" class="tooltip">Enter a value between 0 and 1</span>
-                    <label for="lastmodFilter">Last Modified After:</label>
-                    <input type="date" id="lastmodFilter">
-                    <label for="sortBy">Sort By:</label>
-                    <select id="sortBy">
-                        <option value="url">URL</option>
-                        <option value="priority">Priority</option>
-                        <option value="lastmod">Last Modified</option>
-                    </select>
-                    <button id="applyFiltersBtn">Apply Filters</button>
-                    <button id="resetFiltersBtn">Reset Filters</button>
-                </div>
-            </div>
-            <div id="status">
-                <div id="loading" class="hidden">Loading... <progress id="loadProgress" value="0" max="100"></progress></div>
-                <div id="feedback" class="hidden"></div>
-                <div id="error" class="hidden"></div>
-            </div>
-            <h2>Tree View <span id="matchCount" class="match-count"></span></h2>
-            <div class="tree" id="tree" role="tree" aria-label="Sitemap hierarchy"></div>
-        </section>
-    </main>
-    <footer></footer>
-    <script src="results.js"></script>
-</body>
-</html>
