+++ src/client/index.js
@@ -1,0 +1,8 @@
+import setupInputPage from './pages/InputPage.js';
+import setupResultsPage from './pages/ResultsPage.js';
+
+if (window.location.pathname === '/') {
+  setupInputPage();
+} else if (window.location.pathname === '/results') {
+  setupResultsPage();
+}
